# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type PushTokenRegistrationPlatform = typeof PushTokenRegistrationPlatform[keyof typeof PushTokenRegistrationPlatform];


export const PushTokenRegistrationPlatform = {
  android: 'android',
  ios: 'ios',
} as const;

export interface PushTokenRegistration {
  /** @minLength 1 */
  userId: string;
  /** @minLength 1 */
  token: string;
  platform: PushTokenRegistrationPlatform;
}

export interface MessageNotification {
  /** @minLength 1 */
  recipientUserId: string;
  /** @minLength 1 */
  senderName: string;
  /** @minLength 1 */
  roomId: string;
}

export type CallNotificationMode = typeof CallNotificationMode[keyof typeof CallNotificationMode];


export const CallNotificationMode = {
  audio: 'audio',
  video: 'video',
} as const;

export interface CallNotification {
  /** @minLength 1 */
  recipientUserId: string;
  /** @minLength 1 */
  callerName: string;
  mode: CallNotificationMode;
  /** @minLength 1 */
  callId: string;
}

export interface NotificationStatus {
  ok: boolean;
  messageId?: string;
}