# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: notifications
    description: Firebase Cloud Messaging operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /notifications/register:
    post:
      operationId: registerPushToken
      tags: [notifications]
      summary: Register a device for push notifications
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PushTokenRegistration"
      responses:
        "200":
          description: Token registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationStatus"
        "400":
          description: Invalid registration
        "503":
          description: Push notifications are not configured
  /notifications/send-message:
    post:
      operationId: sendMessageNotification
      tags: [notifications]
      summary: Send a new-message push notification
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MessageNotification"
      responses:
        "200":
          description: Notification sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationStatus"
        "400":
          description: Invalid notification
        "503":
          description: Push notifications are not configured
  /notifications/send-call:
    post:
      operationId: sendCallNotification
      tags: [notifications]
      summary: Send an incoming-call push notification
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CallNotification"
      responses:
        "200":
          description: Notification sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationStatus"
        "400":
          description: Invalid notification
        "503":
          description: Push notifications are not configured
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    PushTokenRegistration:
      type: object
      properties:
        userId:
          type: string
          minLength: 1
        token:
          type: string
          minLength: 1
        platform:
          type: string
          enum: [android, ios]
      required: [userId, token, platform]
    MessageNotification:
      type: object
      properties:
        recipientUserId:
          type: string
          minLength: 1
        senderName:
          type: string
          minLength: 1
        roomId:
          type: string
          minLength: 1
      required: [recipientUserId, senderName, roomId]
    CallNotification:
      type: object
      properties:
        recipientUserId:
          type: string
          minLength: 1
        callerName:
          type: string
          minLength: 1
        mode:
          type: string
          enum: [audio, video]
        callId:
          type: string
          minLength: 1
      required: [recipientUserId, callerName, mode, callId]
    NotificationStatus:
      type: object
      properties:
        ok:
          type: boolean
        messageId:
          type: string
      required: [ok]