# lib/api-zod/src/generated/types/callNotification.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CallNotificationMode } from './callNotificationMode';

export interface CallNotification {
  /** @minLength 1 */
  recipientUserId: string;
  /** @minLength 1 */
  callerName: string;
  mode: CallNotificationMode;
  /** @minLength 1 */
  callId: string;
}