# lib/api-zod/src/generated/types/messageNotification.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface MessageNotification {
  /** @minLength 1 */
  recipientUserId: string;
  /** @minLength 1 */
  senderName: string;
  /** @minLength 1 */
  roomId: string;
}