# lib/api-zod/src/generated/types/pushTokenRegistration.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PushTokenRegistrationPlatform } from './pushTokenRegistrationPlatform';

export interface PushTokenRegistration {
  /** @minLength 1 */
  userId: string;
  /** @minLength 1 */
  token: string;
  platform: PushTokenRegistrationPlatform;
}